// src/generated/models/Oman_qpccategorysectionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Oman_qpccategorysectionsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Oman_qpccategorysectionsstatecode = keyof typeof Oman_qpccategorysectionsstatecode;
export const Oman_qpccategorysectionsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Oman_qpccategorysectionsstatuscode = keyof typeof Oman_qpccategorysectionsstatuscode;

export interface Oman_qpccategorysectionsBase {
  importsequencenumber?: number;
  "oman_CategoryID@odata.bind": string;
  oman_name: string;
  oman_qpccategorysectionid: string;
  "oman_SectionID@odata.bind": string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Oman_qpccategorysectionsstatecode;
  statuscode?: Oman_qpccategorysectionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Oman_qpccategorysections extends Oman_qpccategorysectionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  oman_categoryidname?: string;
  oman_sectionidname?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  oman_categoryid?: object;
  _oman_categoryid_value?: string;
  oman_sectionid?: object;
  _oman_sectionid_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}