// src/generated/models/Oman_qpccategoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Oman_qpccategoriesstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Oman_qpccategoriesstatecode = keyof typeof Oman_qpccategoriesstatecode;
export const Oman_qpccategoriesstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Oman_qpccategoriesstatuscode = keyof typeof Oman_qpccategoriesstatuscode;

export interface Oman_qpccategoriesBase {
  importsequencenumber?: number;
  oman_categoryid: string;
  oman_info?: string;
  oman_name: string;
  oman_qpccategoryid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Oman_qpccategoriesstatecode;
  statuscode?: Oman_qpccategoriesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Oman_qpccategories extends Oman_qpccategoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}