// src/generated/models/Oman_qpcsectionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Oman_qpcsectionsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Oman_qpcsectionsstatecode = keyof typeof Oman_qpcsectionsstatecode;
export const Oman_qpcsectionsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Oman_qpcsectionsstatuscode = keyof typeof Oman_qpcsectionsstatuscode;

export interface Oman_qpcsectionsBase {
  importsequencenumber?: number;
  oman_info?: string;
  oman_name: string;
  oman_qpcsectionid: string;
  oman_sectionid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Oman_qpcsectionsstatecode;
  statuscode?: Oman_qpcsectionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Oman_qpcsections extends Oman_qpcsectionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}